# old-app.go
// Copyright 2016 Google Inc. All rights reserved.
// Use of this source code is governed by the Apache 2.0
// license that can be found in the LICENSE file.

// rugby scores server
package main

//
// import (
// 	"encoding/base64"
// 	"encoding/json"
// 	"fmt"
// 	"log"
// 	"net/http"
//
// 	//"github.com/dominictracey/rugby-scores/model"
// 	"github.com/gorilla/mux"
// 	"google.golang.org/appengine"
// )
//
// func main() {
// 	r := mux.NewRouter()
//
// 	r.Path("/echo").Methods("POST").
// 		HandlerFunc(echoHandler)
//
// 	// pilot Access
// 	r.Path("/pilot").Methods("GET").
// 		HandlerFunc(pilotGetHandler)
//
// 	r.Path("/auth/info/googlejwt").Methods("GET").
// 		HandlerFunc(authInfoHandler)
// 	r.Path("/auth/info/googleidtoken").Methods("GET").
// 		HandlerFunc(authInfoHandler)
// 	r.Path("/auth/info/firebase").Methods("GET", "OPTIONS").
// 		Handler(corsHandler(authInfoHandler))
// 	r.Path("/auth/info/auth0").Methods("GET").
// 		HandlerFunc(authInfoHandler)
//
// 	http.Handle("/", r)
// 	//log.Fatal(http.ListenAndServe(":8000", r))
// 	log.Print("Listening on port 8080")
// 	appengine.Main()
// }
//
// // echoHandler reads a JSON object from the body, and writes it back out.
// func echoHandler(w http.ResponseWriter, r *http.Request) {
// 	var msg interface{}
// 	if err := json.NewDecoder(r.Body).Decode(&msg); err != nil {
// 		if _, ok := err.(*json.SyntaxError); ok {
// 			errorf(w, http.StatusBadRequest, "Body was not valid JSON: %v", err)
// 			return
// 		}
// 		errorf(w, http.StatusInternalServerError, "Could not get body: %v", err)
// 		return
// 	}
//
// 	b, err := json.Marshal(msg)
// 	if err != nil {
// 		errorf(w, http.StatusInternalServerError, "Could not marshal JSON: %v", err)
// 		return
// 	}
//
// 	w.Write(b)
// }
//
// // Pilot represents a drone pilot
// type Pilot struct {
// 	ID       int64
// 	Name     string
// 	Licensed bool
// 	Address  string
// 	Phone    string
// }
//
// // NewPilot returns a new Pilot instance
// func NewPilot() *Pilot {
// 	var i = new(Pilot)
//
// 	//i.Id = "a.assign(metadata.InstanceID)"
// 	i.Name = "Fred Smith"
// 	i.Licensed = true
// 	i.Address = "98 Wallaby Way, Sydney AUS"
// 	i.Phone = "+23 0903 91203"
//
// 	return i
// }
//
// func pilotGetHandler(w http.ResponseWriter, r *http.Request) {
// 	pilot := NewPilot()
// 	pilot.Address = "123 Wallaby Way"
//
// 	b, err := json.Marshal(pilot)
// 	if err != nil {
// 		errorf(w, http.StatusInternalServerError, "Could not marshal JSON: %v", err)
// 		return
// 	}
//
// 	log.Print("call pilot GET")
//
// 	w.Write(b)
// }
//
// // corsHandler wraps a HTTP handler and applies the appropriate responses for Cross-Origin Resource Sharing.
// type corsHandler http.HandlerFunc
//
// func (h corsHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
// 	w.Header().Set("Access-Control-Allow-Origin", "*")
// 	if r.Method == "OPTIONS" {
// 		w.Header().Set("Access-Control-Allow-Headers", "Authorization")
// 		return
// 	}
// 	h(w, r)
// }
//
// // authInfoHandler reads authentication info provided by the Endpoints proxy.
// func authInfoHandler(w http.ResponseWriter, r *http.Request) {
// 	encodedInfo := r.Header.Get("X-Endpoint-API-UserInfo")
// 	if encodedInfo == "" {
// 		w.Write([]byte(`{"id": "anonymous"}`))
// 		return
// 	}
//
// 	b, err := base64.StdEncoding.DecodeString(encodedInfo)
// 	if err != nil {
// 		errorf(w, http.StatusInternalServerError, "Could not decode auth info: %v", err)
// 		return
// 	}
// 	w.Write(b)
// }
//
// // errorf writes a swagger-compliant error response.
// func errorf(w http.ResponseWriter, code int, format string, a ...interface{}) {
// 	var out struct {
// 		Code    int    `json:"code"`
// 		Message string `json:"message"`
// 	}
//
// 	out.Code = code
// 	out.Message = fmt.Sprintf(format, a...)
//
// 	b, err := json.Marshal(out)
// 	if err != nil {
// 		http.Error(w, `{"code": 500, "message": "Could not format JSON for original message."}`, 500)
// 		return
// 	}
//
// 	http.Error(w, string(b), code)
// }